Add player that avoids getting chosen as a target.
The basic idea is that most players have code that looks like this:

    others = Game.world[:players].select { |p| p != self }

This player just defines == so that it's always true for any
PlayerProxy, and so players pretty much skip over it. :) A more
cheatproof way is to compare #object_id, as that way you're not giving
the PlayerProxy object any arguments to play with, so it doesn't know
who is looking at it.

Other than that, it's a pretty braindead attack-the-weakest-guy bot.