Allows classes in player files
This skips adding of non-module descendants of module as players
in engine.rb, so that objects added to the object space
in player files which descend from Module (classes) but are
not actually modules will be skipped.

This is a band-aid fix.

It's looking like some design changes might be nececessary to fix
the root of strange behaviour exhibited by the ObjectSpace based
mechanism for adding players.