Adds new player Jax
Jax is definitely a competitor, but certainly not a dominating force
like dank.  (I'm sure he's a nice guy, but his MMO playing AI is a
total dick.)

I feel that this player adds a lot of value to the game, however, as
he nerfs dank's whorde a bit by casting a confusions spell on it.
(I took advantage of his public method which is used to communicate
targets to make them attack the Dan Knox bot.)  This drops his win
rate about 40% it seems, based of course on the random order in which
players are allowed to act.  Anyway, the nerf is nice, now we all have
a fighting chance!

It turns out my bot plays pretty much just like everyone else.  Nothing
really creative going on aside from a simple rule-based strategy.
(Which I totally thought was the pinnacle of modern MMO tactics till I
realized all the other bots do the same thing.)

He's geared for the endurance game, as he tries to avoid getting too
high leveled right away to avoid the top-player gank-swarm the other
bots tend to form.

Game on!